Restructure origin of props
