+++ packages/ui/src/components/FormField/index.tsx
@@ -1,6 +1,11 @@
-import React, { forwardRef } from "react";
-import { FormFieldProps } from "../../theming/ThemeType";
+import React from "react";
 import { useThemedElement } from "../../theming/useThemedElement";
+
+export type FormFieldProps = {
+    fieldTitle: React.ReactNode;
+    fieldInput: React.ReactNode;
+    fieldErrors?: React.ReactNode;
+};
 
 export const FormField: React.FC<FormFieldProps> = ({ ...props }) => {
     const ThemedFormField = useThemedElement("formField");
