+++ packages/ui/src/themes/html/Button.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { forwardRef } from "react";
-import { ButtonProps } from "../../theming/ThemeType";
+import { ButtonProps } from "../../components/Button";
 
 export const Button = forwardRef<HTMLButtonElement, ButtonProps>(
     ({ children, ...props }, ref) => (
