+++ packages/ui/src/themes/html/RangeSlider.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { forwardRef } from "react";
-import { RangeSliderProps } from "../../theming/ThemeType";
 import * as styles from "./RangeSlider.module.css";
+import { RangeSliderProps } from "../../components/RangeSlider";
 
 export const RangeSlider = forwardRef<HTMLInputElement, RangeSliderProps>(
     ({ children, labels, ...props }, ref) => {
